+++ src/tests/system/libroot/posix/posixtestsuite/conformance/interfaces/mmap/11-1.c
@@ -54,8 +54,8 @@ int main()
   total_size = 3 * page_size;
   size = page_size;
   
-  snprintf(tmpfname, sizeof(tmpfname), "/tmp/pts_mmap_11_1_%d",
-           getpid());
+  snprintf(tmpfname, sizeof(tmpfname), "/tmp/pts_mmap_11_1_%ld",
+           (long)getpid());
   unlink(tmpfname);
   fd = open(tmpfname, O_CREAT | O_RDWR | O_EXCL,
             S_IRUSR | S_IWUSR);
