+++ src/tests/system/libroot/posix/posixtestsuite/conformance/interfaces/mmap/11-2.c
@@ -72,8 +72,8 @@ int main()
   sigaction(SIGBUS, &sa, NULL);
 
   /* Create tmp file */
-  snprintf(tmpfname, sizeof(tmpfname), "/tmp/pts_mmap_11_2_%d",
-           getpid());
+  snprintf(tmpfname, sizeof(tmpfname), "/tmp/pts_mmap_11_2_%ld",
+           (long)getpid());
   unlink(tmpfname);
   fd = open(tmpfname, O_CREAT | O_RDWR | O_EXCL,
             S_IRUSR | S_IWUSR);
