+++ src/tests/system/libroot/posix/posixtestsuite/conformance/interfaces/mmap/11-5.c
@@ -69,8 +69,8 @@ int main()
   /* mmap will create a partial page */
   len = page_size / 2; 
 
-  snprintf(tmpfname, sizeof(tmpfname), "pts_mmap_11_5_%d",
-             getpid());
+  snprintf(tmpfname, sizeof(tmpfname), "pts_mmap_11_5_%ld",
+             (long)getpid());
   child = fork();
   if (child == 0)
   {
