+++ src/tests/system/libroot/posix/posixtestsuite/conformance/interfaces/mmap/12-1.c
@@ -52,8 +52,8 @@ int main()
   total_size = 1024;
   size = total_size;
   
-  snprintf(tmpfname, sizeof(tmpfname), "/tmp/pts_mmap_12_1_%d",
-           getpid());
+  snprintf(tmpfname, sizeof(tmpfname), "/tmp/pts_mmap_12_1_%ld",
+           (long)getpid());
   unlink(tmpfname);
   fd = open(tmpfname, O_CREAT | O_RDWR | O_EXCL,
             S_IRUSR | S_IWUSR);
