+++ src/tests/system/libroot/posix/posixtestsuite/conformance/interfaces/mmap/18-1.c
@@ -75,8 +75,8 @@ int main()
 		return PTS_UNRESOLVED;
   }
   
-  snprintf(tmpfname, sizeof(tmpfname), "pts_mmap_18_1_%d",
-           getpid());
+  snprintf(tmpfname, sizeof(tmpfname), "pts_mmap_18_1_%ld",
+           (long)getpid());
 	
   /* Create shared object */
 	shm_unlink(tmpfname);
