+++ src/tests/system/libroot/posix/posixtestsuite/conformance/interfaces/mmap/21-1.c
@@ -44,8 +44,8 @@ int main()
   total_size = 1024;
   size = total_size;
   
-  snprintf(tmpfname, sizeof(tmpfname), "/tmp/pts_mmap_21_1_%d",
-           getpid());
+  snprintf(tmpfname, sizeof(tmpfname), "/tmp/pts_mmap_21_1_%ld",
+           (long)getpid());
   unlink(tmpfname);
   fd = open(tmpfname, O_CREAT | O_RDWR | O_EXCL,
             S_IRUSR | S_IWUSR);
