+++ src/tests/system/libroot/posix/posixtestsuite/conformance/interfaces/mmap/24-1.c
@@ -51,8 +51,8 @@ int main()
   
   size_t mapped_size = 0; 
 
-  snprintf(tmpfname, sizeof(tmpfname), "pts_mmap_25_1_%d",
-           getpid());
+  snprintf(tmpfname, sizeof(tmpfname), "pts_mmap_25_1_%ld",
+           (long)getpid());
 	
   /* Create shared object */
 	shm_unlink(tmpfname);
