+++ src/tests/system/libroot/posix/posixtestsuite/conformance/interfaces/mmap/27-1.c
@@ -62,8 +62,8 @@ int main()
 #endif
 
   data = (char *) malloc(total_size); 
-  snprintf(tmpfname, sizeof(tmpfname), "/tmp/pts_mmap_27_1_%d",
-           getpid());
+  snprintf(tmpfname, sizeof(tmpfname), "/tmp/pts_mmap_27_1_%ld",
+           (long)getpid());
   unlink(tmpfname);
   fd = open(tmpfname, O_CREAT | O_RDWR | O_EXCL,
             S_IRUSR | S_IWUSR);
