+++ src/tests/system/libroot/posix/posixtestsuite/conformance/interfaces/mmap/6-6.c
@@ -46,8 +46,8 @@ int main()
   int prot;
 
   /* Create the tmp file */
-  snprintf(tmpfname, sizeof(tmpfname), "/tmp/pts_mmap_6_6_%d",
-           getpid());
+  snprintf(tmpfname, sizeof(tmpfname), "/tmp/pts_mmap_6_6_%ld",
+           (long)getpid());
   unlink(tmpfname);
   fd = open(tmpfname, O_CREAT | O_RDWR | O_EXCL,
             S_IRUSR | S_IWUSR);
