+++ src/tests/system/libroot/posix/posixtestsuite/conformance/interfaces/mmap/7-4.c
@@ -60,8 +60,8 @@ int main()
   int exit_stat; 
   
   /* Create shared object */
-  snprintf(tmpfname, sizeof(tmpfname), "pts_mmap_7_4_%d",
-           getpid());
+  snprintf(tmpfname, sizeof(tmpfname), "pts_mmap_7_4_%ld",
+           (long)getpid());
   shm_unlink(tmpfname);
   shm_fd = shm_open(tmpfname, O_RDWR|O_CREAT|O_EXCL, S_IRUSR|S_IWUSR);
   if(shm_fd == -1)
