+++ src/tests/system/libroot/posix/posixtestsuite/conformance/interfaces/sem_post/1-1.c
@@ -32,7 +32,7 @@ int main() {
 	char semname[20];
 	int val;
 
-	sprintf(semname, "/" FUNCTION "_" TEST "_%d", getpid());
+	sprintf(semname, "/" FUNCTION "_" TEST "_%ld", (long)getpid());
 
 	/* Initial value of Semaphore is 0 */
 	mysemp = sem_open(semname, O_CREAT, 0777, 0);
